update make iso install curl check

+++ CP_Source/Tools_Drivers/Make_OSC_Unattended_ISO_Boot/build-igel-iso-os12.sh
@@ -2,7 +2,10 @@
 
 # Fetch and display IGEL OS versions, download, unpack, patch ISO, write to USB
 
-sudo apt install curl -y
+# Install curl if not installed
+if [ ! -f /usr/bin/curl ]; then
+  sudo apt install curl -y
+fi
 
 RELEASE_URL="https://igel-community.github.io/IGEL-Docs-v02/Docs/ReleaseNotes/04-OS12/"
 DOWNLOAD_BASE_URL="https://igeldownloadprod-bydsc8hmbsaegvdy.a01.azurefd.net/files/IGEL_OS_12/OSC"
